Create new folder and file for home page
